+++ packages/orq-rc/RELEASES.md
@@ -8589,3 +8589,13 @@ Based on:
 - [typescript v4.8.0-rc.52] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.52] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.52 - packages/orq-rc
+
+## 2026-04-23 12:21:13
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.53] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.53] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.53 - packages/orq-rc